requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
databases==0.9.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
reportlab==4.2.5
